class name refactoring + saveNewDeviceLocation() obfuscation
