--- app/src/main/java/org/helpapaw/helpapaw/data/repositories/PushNotifications.java
@@ -1,5 +1,0 @@
-package org.helpapaw.helpapaw.data.repositories;
-
-public interface PushNotifications {
-    void registerDeviceForToken();
-}
