+++ app/src/main/java/org/helpapaw/helpapaw/data/repositories/PushNotificationsRepository.java
@@ -1,0 +1,8 @@
+package org.helpapaw.helpapaw.data.repositories;
+
+import android.location.Location;
+
+public interface PushNotificationsRepository {
+    void registerDeviceForToken();
+    void saveNewDeviceLocation(Location location);
+}
