+++ app/src/main/java/org/helpapaw/helpapaw/utils/Injection.java
@@ -1,14 +1,12 @@
 package org.helpapaw.helpapaw.utils;
 
-import android.content.SharedPreferences;
-
 import org.helpapaw.helpapaw.data.repositories.BackendlessCommentRepository;
 import org.helpapaw.helpapaw.data.repositories.BackendlessPhotoRepository;
-import org.helpapaw.helpapaw.data.repositories.BackendlessPushNotifications;
+import org.helpapaw.helpapaw.data.repositories.BackendlessPushNotificationsRepository;
 import org.helpapaw.helpapaw.data.repositories.BackendlessSignalRepository;
 import org.helpapaw.helpapaw.data.repositories.CommentRepository;
 import org.helpapaw.helpapaw.data.repositories.PhotoRepository;
-import org.helpapaw.helpapaw.data.repositories.PushNotifications;
+import org.helpapaw.helpapaw.data.repositories.PushNotificationsRepository;
 import org.helpapaw.helpapaw.data.repositories.SettingsRepository;
 import org.helpapaw.helpapaw.data.repositories.SignalRepository;
 import org.helpapaw.helpapaw.data.user.BackendlessUserManager;
@@ -26,7 +24,7 @@ public class Injection {
     private static PhotoRepository photoRepository;
     private static CommentRepository commentRepository;
     private static SettingsRepository settingsRepository;
-    private static PushNotifications pushNotifications;
+    private static PushNotificationsRepository pushNotificationsRepository;
 
     public synchronized static ImageLoader getImageLoader() {
         if (imageLoader == null) {
@@ -71,11 +69,11 @@ public class Injection {
         return settingsRepository;
     }
 
-    public synchronized static PushNotifications getPushNotificationsInstance() {
-        if (pushNotifications == null) {
-            pushNotifications = new BackendlessPushNotifications();
+    public synchronized static PushNotificationsRepository getPushNotificationsInstance() {
+        if (pushNotificationsRepository == null) {
+            pushNotificationsRepository = new BackendlessPushNotificationsRepository();
         }
 
-        return pushNotifications;
+        return pushNotificationsRepository;
     }
 }
