+++ app/src/main/java/org/helpapaw/helpapaw/utils/services/BackgroundCheckJobService.java
@@ -67,7 +67,7 @@ public class BackgroundCheckJobService extends JobService {
                             //Got last known location. In some rare situations this can be null.
                             if (location != null) {
                                 getSignalsForLastKnownLocation(location, job);
-                                saveNewDeviceLocation(location);
+                                Injection.getPushNotificationsInstance().saveNewDeviceLocation(location);
                             } else {
                                 Log.d(TAG, "got callback but last location is null");
                                 jobFinished(job, true);
@@ -144,67 +144,5 @@ public class BackgroundCheckJobService extends JobService {
                 jobFinished(job, true);
             }
         });
-
-    }
-
-    /*
-     * Queries Backendless for locally saved device-token,
-     * and updates 4 properties on the corresponding db-entry
-     */
-    private void saveNewDeviceLocation(final Location location) {
-
-        // Get local device-token
-        String localToken = Injection.getSettingsRepository().getTokenFromPreferences();
-
-        // Make sure localToken exists
-        if (localToken != null) {
-
-            // Build query
-            String whereClause = "deviceToken = '" + localToken + "'";
-            DataQueryBuilder queryBuilder = DataQueryBuilder.create();
-            queryBuilder.setWhereClause(whereClause);
-
-            Backendless.Data.of("DeviceRegistration").find(queryBuilder,
-                    new AsyncCallback<List<Map>>() {
-                        @Override
-                        public void handleResponse(List<Map> foundDevice) {
-                            // every loaded object from the "DeviceRegistration" table
-                            // is now an individual java.util.Map
-
-                            // Extract 'Map' object from the 'List<Map>'
-                            Map mapFoundDevice = foundDevice.get(0);
-                            try {
-                                mapFoundDevice.put("signalRadius", Injection.getSettingsRepository().getRadius());
-                                mapFoundDevice.put("lastLatitude", location.getLatitude());
-                                mapFoundDevice.put("lastLongitude", location.getLongitude());
-                                mapFoundDevice.put("signalTimeout", Injection.getSettingsRepository().getTimeout());
-                            }
-                            catch (Error e) {
-                                Log.e(TAG, e.getMessage());
-                            }
-
-                            // Save updated object
-                            Backendless.Persistence.save(mapFoundDevice, new AsyncCallback<Map>() {
-                                @Override
-                                public void handleResponse(Map response) {
-                                    Log.d(TAG, "obj updated");
-                                }
-
-                                @Override
-                                public void handleFault(BackendlessFault fault) {
-                                    Log.d(TAG, fault.getMessage());
-                                }
-                            });
-                        }
-
-                        @Override
-                        public void handleFault(BackendlessFault fault) {
-                            // an error has occurred, the error code can be retrieved with fault.getCode()
-                            Log.d(TAG, fault.getMessage());
-                        }
-                    });
-        } else {
-            Log.d(TAG, "localToken is null -or- non-existent");
-        }
     }
 }
